+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0
+
+- Refactored Apis
+- Fix android permission issues
+- Update callback handlers issue
+- Fix add services
+- Some more bug fixes
+
 ## 0.0.1
 
-* Initial version
+- Initial version